+++ debug/site.js
@@ -13,12 +13,13 @@ if (typeof document !== 'undefined') {
                 url: 'http://{s}.gl-api-us-east-1.tilestream.net/v3/mapbox.mapbox-streets-v4/{z}/{x}/{y}.gl.pbf',
                 // urls: ['http://api.tiles.mapbox.com/v3/mapbox.mapbox-streets-v4/{z}/{x}/{y}.vector.pbf'],
                 tileSize: 512,
-                zooms: [0, 2, 3, 4, 5, 6, 7, 8, 10, 12, 13, 14],
+                maxZoom: 14,
+                skipZooms: [1, 9, 11]
             },
             "satellite": {
                 type: 'raster',
                 url: 'http://api.tiles.mapbox.com/v3/aibram.map-vlob92uz/{z}/{x}/{y}.png',
-                zooms: [0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17],
+                maxZoom: 17
             }
         },
         zoom: 15,